Add files via upload
This part use the functions built in the Option_Pricing file. We constructed functions to evaluate European & Binary options prices under the BS formula. Asian options are also evaluated but under Monte Carlo Simulation (this part has to be completed). 

We also compute the greeks of European & Binary options, before creating a portfolio. We can observe the greeks of the new portfolio.